+++ core/components/sprout/docs/changelog.md
@@ -3,6 +3,7 @@
 ## Sprout 0.2.0
 Released on ...
 
+- Add snippet for importing articles
 - Add plugin for generating static HTML
 - Add snippet for processing Markdown
 - Replace Parsedown with League/CommonMark